refactor: flatten @nuxt/webpack package structure
