+++ packages/webpack/src/config/index.js
@@ -1,4 +1,3 @@
 export { default as ClientConfig } from './client'
 export { default as ModernConfig } from './modern'
 export { default as ServerConfig } from './server'
-export { default as PerfLoader } from './utils/perf-loader'
