+++ packages/webpack/src/plugins/vue/client.js
