+++ packages/webpack/src/plugins/vue/modern.js
