+++ packages/webpack/src/plugins/vue/util.js
