+++ packages/webpack/src/plugins/warnfix.js
