+++ packages/webpack/src/utils/perf-loader.js
