+++ packages/webpack/src/utils/style-loader.js
